feat: add colorful login background image
