+++ src/index.tsx
@@ -9,8 +9,8 @@ import MomentUtils from '@date-io/moment'
 import { SnackbarProvider } from 'notistack'
 import * as serviceWorker from './serviceWorker'
 import { SnackbarUtilsConfigurator } from './components/Toast/Toast'
-import Login from './pages/Login/Login'
-import Register from './pages/Register/Register'
+import Login from './pages/Auth/Login'
+import Register from './pages/Auth/Register'
 import Loading from './components/Loading/InstagramLoading'
 import client from './graphql/apolloClient'
 import history from './shared/history'
